shield files that use mpi with if HAVE_MPI
This way the EXPRO and math libraries can be built with out mpi, just not including the files that "include "mpif.h"".
This may not be complete library implemenation but should be exposed upon link time for neo.

Now need to move onto neo and getting it to not need MPI or using the same type of if-deffing.
This will require neo.f90 ->.F90 so it is preprocessed.

Getting this to work with all the make.ext.* files may be tricky.  If NEO runs only serial, then it should not have the
MPI calls.