Cleaning up write_distributed_complex_h5
This cleans up the naming scheme to avoid redundant logic.
This could potentially be cleaned up further to make it common
between the real and complex, but we can worry about that later.
We've also added grouping levels based on feedback from Mark
and discussions on what we want visit to look like when we pull
in a file.  For the python scripting, this should simplify the
overall imports as well, although we will need to drill down a level.