Fixed the writing to hdf5 files.
It is necessary to have all strings "TRIM"ed when passing to hdf5.
I have added a drift or gk prefix to the electrons to let the user know which electron method is being used.
This should only write out evolved quantities,ie. in the case electron_method=3, then only gk_electrons are dumped since
ions are addiabatic (ETG simulations).