aggregate time file has wiki dimensions.
out.gyro.timedata.h5 contains data that corresponds to the wiki for gyro output.
The tricky thing was making sure to get the n-D arrays written out with the same dimensional ordering as the wiki.
This was a game of passing a fortran nd array to C (hdf5) such that C (hdf5) would write it out in the fortran ordering.
It required reordering before passing, and keeping track of all the extent of existing data when adding new data from
the new timestep.