Timer causing issues with hdf5 out.
When both output formats are written (IO_METHOD>1) the timer
variable cpu(:) gets clobbered and can become zero.  This is a problem because
cpu(cpu_maxindx) is in a denominator for "frac" and thus this seems to hang some
simulations. So, the timer in the hfd5 timedata subroutine is commented out
until this is sorted out.