making 1 file for grid dump
This is easier to maintain and allows for
easier appending vs. overwriting.
It is possible to write to file each of the h5 grids, exclusively.
This is going to have to error shielded for all the
possible combinations.  Might want to dump all three
grids every time even if we are not dumping data each time.

REVISIT.