Mesh seperation working.
So now, all spatial R,Z, toroidal mesh info is in
the file gyroMesh.h5, and all the hdf5 files that have spatial
defined quantities, have hdf5 Groups that have data sets which
point to this file and the proper group that contains the
spatial mesh.  These meshes are poloidal, 3d and the wedge.